feature: Update Splat.Autofac to work with Autofac 5.0 (#465)
Raises Autofac version to 5+, and deprecates or removes method that mutate the container (since that is supposed to be immutable from now on).
Register and HasRegister are to be used only by the ReactiveUI initialization procedure, which modifies the container by creating a child scope for each service registration.
Thanks to "skaman" for coming up with the workaround.